requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
twilio==8.10.0
anthropic
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai
packaging>=23.0
requests>=2.31.0
pytz>=2025.1